SOF-371: Refactor duplicate usecases for collector name and title
