+++ app/src/main/java/com/vci/vectorcamapp/core/domain/util/collector/CollectorValidationError.kt
@@ -1,8 +1,8 @@
-package com.vci.vectorcamapp.registration.domain.util
+package com.vci.vectorcamapp.core.domain.util.collector
 
 import com.vci.vectorcamapp.core.domain.util.Error
 
-enum class RegistrationValidationError : Error {
+enum class CollectorValidationError : Error {
     BLANK_COLLECTOR_TITLE,
     BLANK_COLLECTOR_NAME,
 }
