+++ app/src/main/java/com/vci/vectorcamapp/core/presentation/util/error/ErrorExtensions.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.vci.vectorcamapp.R
 import com.vci.vectorcamapp.complete_session.details.domain.util.CompleteSessionDetailsError
 import com.vci.vectorcamapp.core.domain.util.Error
+import com.vci.vectorcamapp.core.domain.util.collector.CollectorValidationError
 import com.vci.vectorcamapp.core.domain.util.network.NetworkError
 import com.vci.vectorcamapp.core.domain.util.room.RoomDbError
 import com.vci.vectorcamapp.imaging.domain.util.ImagingError
@@ -13,9 +14,7 @@ import com.vci.vectorcamapp.registration.domain.util.RegistrationError
 import com.vci.vectorcamapp.intake.domain.util.FormValidationError
 import com.vci.vectorcamapp.intake.domain.util.IntakeError
 import com.vci.vectorcamapp.landing.domain.util.LandingError
-import com.vci.vectorcamapp.registration.domain.util.RegistrationValidationError
 import com.vci.vectorcamapp.settings.domain.util.SettingsError
-import com.vci.vectorcamapp.settings.domain.util.SettingsValidationError
 
 fun Error.toString(context: Context): String {
     val resId = when (this) {
@@ -107,19 +106,14 @@ fun Error.toString(context: Context): String {
             FormValidationError.BLANK_SPECIMEN_CONDITION -> R.string.form_validation_error_blank_specimen_condition
         }
 
-        is RegistrationValidationError -> when (this) {
-            RegistrationValidationError.BLANK_COLLECTOR_TITLE -> R.string.registration_validation_error_blank_collector_title
-            RegistrationValidationError.BLANK_COLLECTOR_NAME -> R.string.registration_validation_error_blank_collector_name
+        is CollectorValidationError -> when (this) {
+            CollectorValidationError.BLANK_COLLECTOR_TITLE -> R.string.collector_validation_error_blank_collector_title
+            CollectorValidationError.BLANK_COLLECTOR_NAME -> R.string.collector_validation_error_blank_collector_name
         }
 
         is SettingsError -> when (this) {
             SettingsError.COLLECTOR_SAVE_FAILED -> R.string.settings_error_collector_save_failed
             SettingsError.COLLECTOR_DELETION_FAILED -> R.string.settings_error_collector_deletion_failed
-        }
-
-        is SettingsValidationError -> when (this) {
-            SettingsValidationError.BLANK_COLLECTOR_TITLE -> R.string.settings_validation_error_blank_collector_title
-            SettingsValidationError.BLANK_COLLECTOR_NAME -> R.string.settings_validation_error_blank_collector_name
         }
 
         else -> R.string.error_fallback
