--- app/src/main/java/com/vci/vectorcamapp/registration/domain/use_cases/RegistrationValidationUseCases.kt
@@ -1,8 +1,0 @@
-package com.vci.vectorcamapp.registration.domain.use_cases
-
-import javax.inject.Inject
-
-data class RegistrationValidationUseCases @Inject constructor(
-    val validateCollectorTitle: ValidateCollectorTitleUseCase,
-    val validateCollectorName: ValidateCollectorNameUseCase,
-)
