+++ app/src/main/java/com/vci/vectorcamapp/registration/presentation/model/RegistrationErrors.kt
@@ -1,8 +1,8 @@
 package com.vci.vectorcamapp.registration.presentation.model
 
-import com.vci.vectorcamapp.registration.domain.util.RegistrationValidationError
+import com.vci.vectorcamapp.core.domain.util.collector.CollectorValidationError
 
 data class RegistrationErrors(
-    val collectorName: RegistrationValidationError? = null,
-    val collectorTitle: RegistrationValidationError? = null
+    val collectorName: CollectorValidationError? = null,
+    val collectorTitle: CollectorValidationError? = null
 )
