+++ app/src/main/java/com/vci/vectorcamapp/settings/presentation/SettingsViewModel.kt
@@ -2,14 +2,14 @@ package com.vci.vectorcamapp.settings.presentation
 
 import androidx.lifecycle.viewModelScope
 import com.vci.vectorcamapp.core.domain.cache.DeviceCache
+import com.vci.vectorcamapp.core.domain.collector.use_cases.CollectorValidationUseCases
 import com.vci.vectorcamapp.core.domain.model.Collector
 import com.vci.vectorcamapp.core.domain.model.enums.SessionType
 import com.vci.vectorcamapp.core.domain.repository.CollectorRepository
 import com.vci.vectorcamapp.core.domain.repository.ProgramRepository
-import com.vci.vectorcamapp.core.domain.util.errorOrNull
 import com.vci.vectorcamapp.core.domain.util.Result
+import com.vci.vectorcamapp.core.domain.util.errorOrNull
 import com.vci.vectorcamapp.core.presentation.CoreViewModel
-import com.vci.vectorcamapp.settings.domain.use_cases.SettingsValidationUseCases
 import com.vci.vectorcamapp.settings.domain.util.SettingsError
 import com.vci.vectorcamapp.settings.presentation.model.SettingsErrors
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -30,7 +30,7 @@ class SettingsViewModel @Inject constructor (
     private val deviceCache: DeviceCache,
     private val programRepository: ProgramRepository,
     private val collectorRepository: CollectorRepository,
-    private val settingsValidationUseCases: SettingsValidationUseCases
+    private val collectorValidationUseCases: CollectorValidationUseCases
 ): CoreViewModel(){
 
     private val _collectors = collectorRepository.observeAllCollectors()
@@ -111,8 +111,8 @@ class SettingsViewModel @Inject constructor (
                 SettingsAction.SaveCollector -> {
                     val collector = state.value.selectedCollector ?: return@launch
 
-                    val nameValidationResult = settingsValidationUseCases.validateCollectorName(collector.name)
-                    val titleValidationResult = settingsValidationUseCases.validateCollectorTitle(collector.title)
+                    val nameValidationResult = collectorValidationUseCases.validateCollectorName(collector.name)
+                    val titleValidationResult = collectorValidationUseCases.validateCollectorTitle(collector.title)
 
                     _state.update { currentState ->
                         currentState.copy(
