+++ app/src/main/res/values/strings.xml
@@ -155,10 +155,10 @@
     <string name="registration_error_unknown_error">An unknown error occurred. Please try again or contact support if the issue persists.</string>
 
     <!-- Registration validation error: Blank collector title -->
-    <string name="registration_validation_error_blank_collector_title">Collector title is required.</string>
+    <string name="collector_validation_error_blank_collector_title">Collector title is required.</string>
 
     <!-- Registration validation error: Blank collector name -->
-    <string name="registration_validation_error_blank_collector_name">Collector name is required.</string>
+    <string name="collector_validation_error_blank_collector_name">Collector name is required.</string>
 
     <!-- Incomplete session error: Session not found -->
     <string name="incomplete_session_error_session_not_found">Session not found. Please try again.</string>
@@ -189,12 +189,6 @@
 
     <!-- Settings error: Collector deletion failed -->
     <string name="settings_error_collector_deletion_failed">Could not delete collector. Please try again.</string>
-
-    <!-- Settings validation error: Blank collector title -->
-    <string name="settings_validation_error_blank_collector_title">Collector title is required.</string>
-
-    <!-- Settings validation error: Blank collector name -->
-    <string name="settings_validation_error_blank_collector_name">Collector name is required.</string>
 
     <!-- Image Upload Status: Upload completed -->
     <string name="upload_status_completed">Completed</string>
